requirements.txt
opencv-python
easyocr[gpu]
groq
langdetect
numpy
# paddleocr
pygame
pyttsx3
requests
#torch
#torchvision  
#torchaudio
Pillow
matplotlib
accelerate
transformers
bitsandbytes
safetensors
sentencepiece
protobuf
ollama
  


# SESAME
huggingface_hub
moshi
torchtune
torchao
silentcipher @ git+https://github.com/SesameAILabs/silentcipher@master

# Google Gemini
google-genai


# Frontend and Backend 
fastapi
uvicorn
python-multipart
python-dotenv
pillow
elevenlabs
pathlib
asyncio